v1.73.20_ThrallsBeta
The core mechanics of the Resurgent xenotype (Mechanics of walking corpses) have been completely redesigned. The mechlink is no longer used for control, and corpses must be raised manually. The corpse limit has also been completely redone; now, to increase the limit, you need to increase the number of resurgent pawns in the colony.

The new mechanic is called Thralls. Thralls are specific xenotypes with strict limitations and their own strengths and weaknesses. Thralls are created from corpses using the "Thrall Master" gene, the player can raise an unlimited number of thralls, but their lifespan is limited to 15 days, to prolong it, thralls must regularly absorb the resurgent cells of resurgent pawns.